# docs/swagger.yaml
definitions:
  authz-service_internal_model.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  authz-service_internal_model.CheckResult:
    properties:
      allowed:
        type: boolean
      source:
        type: string
    type: object
  authz-service_internal_model.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/authz-service_internal_model.APIError'
    type: object
...
  authz-service_internal_model.PermissionEntry:
    properties:
      action:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      scope:
        additionalProperties:
          type: string
        type: object
      source:
        type: string
    type: object
  authz-service_internal_model.ResourcesResult:
    properties:
      resource_ids:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /query/v1/check:
    get:
      description: |-
        Check if a user is authorized to perform an action on a specific resource.
        Resolution order: DENY override > ALLOW override > role check > default deny.
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      - description: Action name (e.g. read, write)
        in: query
        name: action
        required: true
        type: string
      - description: Resource type (e.g. project, repo)
        in: query
        name: resource_type
        required: true
        type: string
      - description: Resource identifier
        in: query
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authz-service_internal_model.CheckResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      summary: Check authorization
      tags:
      - query
  /query/v1/permissions:
    get:
      description: List the effective permissions for a user, optionally filtered
        by resource type and resource ID.
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      - description: Filter by resource type
        in: query
        name: resource_type
        type: string
      - description: Filter by resource ID
        in: query
        name: resource_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/authz-service_internal_model.PermissionEntry'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      summary: List effective permissions
      tags:
      - query
  /query/v1/resources:
    get:
      description: Return all resource IDs that a user is authorized to perform an
        action on.
      parameters:
      - description: User UUID
        in: query
        name: user_id
        required: true
        type: string
      - description: Action name
        in: query
        name: action
        required: true
        type: string
      - description: Resource type
        in: query
        name: resource_type
        required: true
        type: string
      - description: JSON-encoded scope filter object
        in: query
        name: scope_filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authz-service_internal_model.ResourcesResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      summary: List authorized resources
      tags:
      - query
  /query/v1/resources/filter:
    post:
      consumes:
      - application/json
      description: Given a list of candidate resource IDs, return only those the user
        is authorized for.
      parameters:
      - description: Filter request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/authz-service_internal_model.ErrorResponse'
      summary: Filter authorized resource IDs
      tags:
      - query
...
swagger: "2.0"